+++ components/MainNavbar/index.js
@@ -1,21 +1,37 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
 import cn from "classnames";
+import useTranslation from "next-translate/useTranslation";
 
 import Socials from "components/Socials";
 
 import Logo from "public/logo.svg";
 import style from "./style.module.scss";
 
-import { routes } from "./constants";
-
 export default function MainNavbar() {
   const { pathname } = useRouter();
+  const { t } = useTranslation("common");
+
+  const routes = [
+    {
+      route: "/",
+      label: t`navbar.home`,
+    },
+    {
+      route: "/blog",
+      label: t`navbar.blog`,
+    },
+    {
+      route: "/recursos",
+      label: t`navbar.resources`,
+    },
+  ];
+
   return (
     <nav className={style.MainNavbar}>
       <Logo className={style.logo} />
       <ul className={style.links}>
-        {routes?.map(({ route, es }, i) => (
+        {routes?.map(({ route, label }, i) => (
           <li key={i}>
             <Link href={route}>
               <a
@@ -26,7 +42,7 @@ export default function MainNavbar() {
                       : pathname.startsWith(route),
                 })}
               >
-                {es}
+                {label}
               </a>
             </Link>
           </li>